refactor(article): extracts series item
