+++ src/v2/Apps/Article/Components/ArticleSeries.tsx
@@ -2,13 +2,10 @@ import {
   Box,
   ChevronIcon,
   Column,
-  Flex,
   GridColumns,
   Join,
-  ResponsiveBox,
   Spacer,
   Text,
-  Image,
   HTML,
   Separator,
 } from "@artsy/palette"
@@ -20,7 +17,7 @@ import { RouterLink } from "v2/System/Router/RouterLink"
 import { ArticleSeries_article } from "v2/__generated__/ArticleSeries_article.graphql"
 import { ArticleAd } from "./ArticleAd"
 import { ArticleSponsorFragmentContainer } from "./ArticleSponsor"
-import styled from "styled-components"
+import { ArticleSeriesItemFragmentContainer } from "./ArticleSeriesItem"
 
 interface ArticleSeriesProps {
   article: ArticleSeries_article
@@ -63,93 +60,11 @@ const ArticleSeries: FC<ArticleSeriesProps> = ({ article }) => {
 
       <Join separator={<Spacer mt={4} />}>
         {article.relatedArticles.map(relatedArticle => {
-          const image = relatedArticle.thumbnailImage?.display
-
           return (
-            <RouterLink
-              to={relatedArticle.href}
-              display="block"
-              textDecoration="none"
-            >
-              <GridColumns
-                key={relatedArticle.internalID}
-                border="1px solid"
-                borderColor="black100"
-                gridRowGap={2}
-                p={[2, 4]}
-              >
-                <Column span={6} order={[1, 0]}>
-                  <Flex
-                    flexDirection="column"
-                    justifyContent="space-between"
-                    height="100%"
-                  >
-                    <Box>
-                      <Text variant="xs" textTransform="uppercase" mb={0.5}>
-                        {article.title}
-                      </Text>
-
-                      <Text variant="xl" mb={2}>
-                        {relatedArticle.thumbnailTitle ?? relatedArticle.title}
-                      </Text>
-
-                      <Text variant="lg">{relatedArticle.description}</Text>
-                    </Box>
-
-                    <Flex mt={4}>
-                      <Text variant="xs" textTransform="uppercase">
-                        {relatedArticle.byline}
-                      </Text>
-
-                      <Text variant="xs" textTransform="uppercase" ml={2}>
-                        {relatedArticle.publishedAt}
-                      </Text>
-                    </Flex>
-                  </Flex>
-                </Column>
-
-                <Column span={6}>
-                  <ResponsiveBox
-                    position="relative"
-                    aspectWidth={869}
-                    aspectHeight={580}
-                    maxWidth="100%"
-                    bg="black30"
-                  >
-                    {image && (
-                      <Image
-                        src={image.src}
-                        srcSet={image.srcSet}
-                        width="100%"
-                        height="100%"
-                        lazyLoad
-                      />
-                    )}
-
-                    {relatedArticle.media && (
-                      <>
-                        <Play
-                          position="absolute"
-                          bottom={2}
-                          left={2}
-                          color="white100"
-                        />
-
-                        <Text
-                          variant="xs"
-                          color="white100"
-                          position="absolute"
-                          bottom={1}
-                          right={2}
-                        >
-                          {relatedArticle.media.duration}
-                        </Text>
-                      </>
-                    )}
-                  </ResponsiveBox>
-                </Column>
-              </GridColumns>
-            </RouterLink>
+            <ArticleSeriesItemFragmentContainer
+              key={relatedArticle.internalID}
+              article={relatedArticle}
+            />
           )
         })}
 
@@ -214,36 +129,10 @@ export const ArticleSeriesFragmentContainer = createFragmentContainer(
           ...ArticleSponsor_sponsor
         }
         relatedArticles {
+          ...ArticleSeriesItem_article
           internalID
-          href
-          title
-          thumbnailTitle
-          byline
-          description
-          publishedAt(format: "MMM DD, YYYY")
-          thumbnailImage {
-            # 3:2 aspect ratio
-            display: cropped(width: 869, height: 580) {
-              src
-              srcSet
-            }
-          }
-          media {
-            duration
-          }
         }
       }
     `,
   }
 )
-
-const Play = styled(Box)`
-  &:after {
-    content: "";
-    display: block;
-    color: currentColor;
-    border-top: 15px solid transparent;
-    border-bottom: 15px solid transparent;
-    border-left: 30px solid;
-  }
-`
